% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperDynNetpar.R
\name{helperDynNetpar}
\alias{helperDynNetpar}
\title{Helper function to give initial model parameters}
\usage{
helperDynNetpar(
  structural.model,
  measurement.model,
  Time,
  Tentry = "Tentry",
  Event = "Event",
  StatusEvent = "StatusEvent",
  basehaz = NULL,
  assocT = NULL,
  subject,
  data
)
}
\arguments{
\item{structural.model}{a list of 7 arguments used to specify the structural model: \describe{
   \item{\code{fixed.LP0}}{a one-sided linear formula object for specifying the 
   fixed effects in the submodel for the baseline level of processes. Note that 
   there is no need to specify a random effect model for the baseline level of 
   processes as we systematically set a process-specific random intercept (with 
   variance fixed to 1 for identifiability purpose). For identifiability purposes, 
   the mean intercepts are fixed to 0 (not estimated).}

   \item{\code{fixed.DeltaLP}}{a two-sided linear formula object for 
   specifying the response outcomes (one the left part of ~ symbol) 
   and the covariates with fixed-effects (on the right part of ~ symbol) 
   in the submodel for change over time of latent processes.}

   \item{\code{random.DeltaLP}}{a one-sided linear formula object 
   for specifying the random effects in the submodel for change 
   over time of latent processes.}

   \item{\code{trans.matrix}}{a one-sided linear formula object 
   for specifying a model for elements of the transition matrix, 
   which captures the temporal influences between latent processes.}

   \item{\code{fixed.survival}}{a one-sided linear formula object 
   for specifying the covariates in the survival sub-model. In competing 
   risks model, the specification for the two events should be separated 
   by the "|" symbol.}

   \item{\code{interactionY.survival}}{a one-sided linear formula 
   object for specifying the covariates in interaction with the 
   dynamics of the latent processes, in the survival sub-model. 
   In competing risks model, the specification for the two events 
   should be separated by the "|" symbol. Only additional terms 
   should be included (No "*" symbol). Covariates in 
   \code{interactionY.survival} should also be included in \code{fixed.survival.}}

   \item{\code{delta.time}}{indicates the discretisation step to be used for latent processes}
   }}

\item{measurement.model}{is a list of arguments detailed below used to specify the measurement model: \describe{

\item{\code{link}}{ indicates the link functions to be used to transform the outcomes. It takes values in "linear" for a linear transformation and "n-type-d" for a I-splines transformation where "n" indicates the number of nodes, "type" (which takes values in "quant", "manual", "equi") indicates where the nodes are placed, and "d" indicates the degree of the I-splines.}

\item{\code{knots}}{ argument indicates if necessary the place of knots (when placed manually with "manual"), default value is NULL}
}}

\item{Time}{indicates the name of the covariate representing the time}

\item{Tentry}{name of the variable of entry time}

\item{Event}{name of the variable of event time}

\item{StatusEvent}{name of the variable of event status}

\item{basehaz}{type of baseline hazard function}

\item{subject}{indicates the name of the covariate representing the grouping structure}

\item{data}{indicates the data frame containing all the variables for estimating the model.}

\item{parameters}{a list of 3 arguments about parameters of the models (e.g., initial parameters, parameters one would like to fix, etc.) possibly created with \link{helperDynNetpar}: \describe{
\item{\code{paras.ini}}{ indicates initial values for parameters, default values is NULL.}
\item{\code{Fixed.para.indix}}{ indicates the positions of parameters to be constrained.}

\item{\code{Fixed.para.values}}{ indicates the values associated to the index of parameters to be constrained. }
}
  \item{\code{assocT}}{ Specifies the type of association between the time-to-event(s) and the latent process(es). Values include "r.intercept" for random intercept, "r.slope" for random slope, "r.intercept/slope" for both random intercept and slope, "c.value" for current value. }}

\item{cholesky}{logical indicating if the variance covariance matrix is parameterized using the cholesky (TRUE by default) or the correlation (FALSE)}
}
\value{
object of class DynNetinit
}
\description{
Helper function to give initial model parameters
}
