% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametre2.R, R/Parametre3.R
\name{Parametre2}
\alias{Parametre2}
\title{Function to initialize parameters k in multivariate DynNet model}
\usage{
Parametre2(
  K,
  nD,
  nL,
  mapping.to.LP,
  mapping.to.LP2,
  vec_ncol_x0n,
  n_col_x,
  nb_RE,
  stochErr = FALSE,
  indexparaFixeUser = NULL,
  paraFixeUser = NULL,
  L = 1,
  paras.ini,
  ncolMod.MatrixY,
  link,
  npara_k,
  Survdata = NULL,
  basehaz = NULL,
  knots_surv = NULL,
  assoc = NULL,
  truncation = F,
  data,
  outcomes,
  df,
  nE = 0,
  np_surv = 0,
  fixed.survival.models = NULL,
  interactionY.survival.models = NULL,
  nYsurv = 0,
  data_F = data_F
)

Parametre2(
  K,
  nD,
  nL,
  mapping.to.LP,
  mapping.to.LP2,
  vec_ncol_x0n,
  n_col_x,
  nb_RE,
  stochErr = FALSE,
  indexparaFixeUser = NULL,
  paraFixeUser = NULL,
  L = 1,
  paras.ini,
  ncolMod.MatrixY,
  link,
  npara_k,
  Survdata = NULL,
  basehaz = NULL,
  knots_surv = NULL,
  assoc = NULL,
  truncation = F,
  data,
  outcomes,
  df,
  nE = 0,
  np_surv = 0,
  fixed.survival.models = NULL,
  interactionY.survival.models = NULL,
  nYsurv = 0,
  data_F = data_F
)
}
\arguments{
\item{K}{number of the markers}

\item{nD}{number of the latent processes}

\item{nL}{number of exogeneous latent processes}

\item{mapping.to.LP}{indicates which outcome measured which exogenous latent process, it is a mapping table between
outcomes and enogenous latents processes (only used for structural model)}

\item{vec_ncol_x0n}{vector of number of columns of model.matrix for baseline's submodel}

\item{n_col_x}{number of overall columns of model.matrix for change's submodel}

\item{nb_RE}{number of random effects}

\item{stochErr}{indicates if the structural model contain stochastique components}

\item{indexparaFixeUser}{position of parameters to be constrained}

\item{paraFixeUser}{values associated to the index of parameters to be constrained}

\item{L}{number of columns of model.matrix for temporal infuences model}

\item{paras.ini}{initial values for parameters, default values is NULL}

\item{ncolMod.MatrixY}{vector of number of columns of model.matrix for transformation submodel}

\item{link}{indicates link used to transform outcome}

\item{npara_k}{marker-specific number of transformation parameters}

\item{Survdata}{dataset for survival model}

\item{basehaz}{type of baseline hazard function}

\item{knots_surv}{knots for splines in baseline hazard (to develop)}

\item{assoc}{specification of association between longitudinal and survival models}

\item{truncation}{boolean for delayed entry}

\item{data}{dataset for longitudinal model}

\item{outcomes}{names of the outcomes}

\item{df}{degree of freedom for link==splines}

\item{nE}{number of survival events}

\item{np_surv}{cause-specific number of fixed effects in survival models}

\item{fixed.survival.models}{specification of survival models (without interactions)}

\item{interactionY.survival.models}{specification of interactions in survival models}

\item{nYsurv}{number of fixed effects in survival model}
}
\value{
a list

a list
}
\description{
Function to initialize parameters k in multivariate DynNet model

Function to initialize parameters k in multivariate DynNet model
}
