% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parskeleton.R
\name{parskeleton}
\alias{parskeleton}
\title{Prepare skeleton object for initial values}
\usage{
parskeleton(
  K,
  nD,
  nL,
  mapping.to.LP,
  mapping.to.LP2,
  vec_ncol_x0n,
  n_col_x,
  nb_RE,
  stochErr = FALSE,
  indexparaFixeUser = NULL,
  paraFixeUser = NULL,
  L = 1,
  paras.ini,
  ncolMod.MatrixY,
  link,
  npara_k,
  Survdata = NULL,
  basehaz = NULL,
  knots_surv = NULL,
  assoc = NULL,
  truncation = F,
  data,
  outcomes,
  df,
  nE = 0,
  np_surv = 0,
  fixed.survival.models = NULL,
  interactionY.survival.models = NULL,
  nYsurv = 0
)
}
\arguments{
\item{K}{}

\item{nD}{}

\item{nL}{}

\item{mapping.to.LP}{}

\item{mapping.to.LP2}{}

\item{vec_ncol_x0n}{}

\item{n_col_x}{}

\item{nb_RE}{}

\item{stochErr}{}

\item{indexparaFixeUser}{}

\item{paraFixeUser}{}

\item{L}{}

\item{paras.ini}{}

\item{ncolMod.MatrixY}{}

\item{link}{}

\item{npara_k}{}

\item{Survdata}{}

\item{basehaz}{}

\item{knots_surv}{}

\item{assoc}{}

\item{truncation}{}

\item{data}{}

\item{outcomes}{}

\item{df}{}

\item{nE}{}

\item{np_surv}{}

\item{fixed.survival.models}{}

\item{interactionY.survival.models}{}

\item{nYsurv}{}
}
\value{
parameters object
}
\description{
Prepare skeleton object for initial values
}
